Add distinctipy
I stumbled across https://github.com/matplotlib/matplotlib/issues/22033 requesting larger qualitative colormaps and saw my package distinctipy (which can be used for this purpose) mentioned in the comments. I'd love for it to be included in the 3rd party list and happy to make any modifications needed for that to happen.