Add explicit name attribute to all packages
This allows using whitespace and arbitrary capitalization, that
need not match how the repository is spelled.

I also think name is so fundamental that it must be mandatory.
This will also save us from trying to figure out the name from
a repository location, which will become more complicated when
we accept non-GitHub repos.