Simplify build script
There appears to be no need to write `all.yml` as nothing outside the
build script uses it. The effect of using it is to re-order items and
flatten the outer dictionary into list.

But we can just do all that work when reading, and the sorting in the
template.